Notes on bad data.
I'm getting bad data from the API I think.

EG, ROBG on the 12th August currently nearly 100x
lower than it should be.

Tried changing the history-period to see if
a different cache somewhere would make it
better but it didn't.

Some of the bad data from the API seemed to get
repaired as I investigated, but some bad data
remains.

Hopefully if it's getting back-fixed we should
overwrite all the bad-data in my own cache
over time. Extend data-grab to 3 months coz
it appears to be costless to me, probably is
nearly costless to them, and hopefully gives
them longer to fix the bad data if they get
bad data sometimes and then fix it later.