There was an exception so I caught it.
I have no idea why I suddenly started getting an
exception for WISH in this code, and I'm too drunk
by far to dig into the details.

So I caught the exception, which I assume is an
error in the source-data, and refused to fail
like it was failing.

And now I actually have a summary for tonight.
Which is surely missing many alerts and has
many false positives but, hey ho. best we can
do for now.